+++ changelog.md
@@ -20,6 +20,7 @@
 - /lfg does not ping role when used inside threads
 - /lfg claims the thread it is used in when possible
 - /lfg does not claim a thread if the role specified belongs to a different forum channel (you can't create a forum specific game outside of that forum)
+- hid confusing survey choice to send results to tournament or first stage when tournament doesn't use stages
 
 ## [1.13.4] - 2026-8-31 Bot Evolution